reformat code, next step: presentation
