+++ README.md
@@ -1,11 +1,43 @@
-# proyecto-algo2
+# Proyecto de Algoritmos y Estructura de Datos II - UNCuyo 2024 📚
 
-Proyecto Algoritmos2 - Base de Datos
+## Integrantes del Equipo 👥
 
-Francisco Aragona
+- Francisco Aragona
+- Renzo Davila
+- Facundo Lella
+- Adriano Fabris
 
-Renzo Davila
+## Descripción del Proyecto 📜
 
-Facundo Lella
+Desarrollar una aplicación para manejar un conjunto de documentos utilizando una
+estructura de datos previamente definida en la misma aplicación.
 
-Adriano Fabris
+Desarrollar mecanismos eficientes de consultas sobre los documentos de la base de
+datos.
+
+## Estructura del Repositorio 📁
+
+- **source/**: Código fuente del proyecto.
+- **categorization/**: Funciones auxiliares.
+- **pickles/**: Dumping y opening de archivos.
+- **pdfs/**: Para casos de prueba.
+
+## Tecnologías Utilizadas 🛠️
+
+- Lenguaje de programación: Python3
+- Librerías: Pickle
+
+## Uso y Ejecución 🚀
+
+1. Ejecutar creación de base de datos:
+    ```bash
+    python3 document_db.py -create <local_path>
+    ```
+2. Ejecutar la búsqueda de documentos similares:
+    ```bash
+    python3 document_db.py -search <text>
+
+---
+
+¡Gracias por la visita! 🌟
+
